move kafka producer to a more generic directory
