+++ udata/kafka/__init__.py
@@ -1,0 +1,8 @@
+from enum import Enum
+from .producer import produce, KafkaProducerSingleton
+
+
+class KafkaMessageType(Enum):
+    INDEX = 'index'
+    REINDEX = 'reindex'
+    UNINDEX = 'unindex'
